INTRODUCE SUBMODULES
I want to build a new client for my game using Apache Wicket. :) So I split the code into "twothousand-core" for the game stuff and "twothousand-web" for Apache Wicket stuff. There is not much happening at the moment, just a single page not even displaying the Playfield yet, but I have everything in place to start.

To get some java8 features I introduced a new functional interface "PlayfieldControlParse" to extract the code to get player input for different clients. I think it's useful to keep this out of the actual control code, and I feel like I have to add more of these functional interfaces when I start adding the controls and the displaying of the Playfield in my web client.

Usually I'd create a new git branch for this kind of change, but for better reviewability I'll just keep it in master. You can still run the Game and just ignore the submodule "twothousand-web" for the moment.