WATCH OUT - this change renders most of the unit tests useless, because a random tile is now automatically added after each move of the playfield. :(
The good part is that the web client now displays a playfield and you can start moving stuff around. The bad thing is that this removes most of the unit tests. There is a random part inside the playfield now that I can not yet override with static values in unit tests. Usually I'd continue to refactor locally until this situation is solved, maybe would throw away the solution when I discover that this is a dead end.

I add this commit for purposes of reviewability only.

Time is up for today, so I'll let this stand here and continue or throw away the last change when I continue.