Try to add more java8 features, starting with Lambdas.
I will try to replace most of the nested for loops for the tile handling with a lambda. That way I can reuse most of the nested for loops and replace the custom logic executed on each tile with a lambda.

Should be useful for the ASCII printing, the score calculation and probably the movement.